backend/requirements.txt
# aiohappyeyeballs==2.6.1
# aiohttp==3.13.0
# aiosignal==1.4.0
# annotated-types==0.7.0
# anyio==4.11.0
# attrs==25.4.0
# backoff==2.2.1
# bcrypt==5.0.0
# build==1.3.0
# cachetools==6.2.0
# certifi==2025.10.5
# charset-normalizer==3.4.3
# chromadb==1.1.1
# click==8.3.0
# colorama==0.4.6
# coloredlogs==15.0.1
# dataclasses-json==0.6.7
# distro==1.9.0
# durationpy==0.10
# edge-tts==7.2.3
# fastapi==0.118.0
# filelock==3.19.1
# filetype==1.2.0
# flatbuffers==25.9.23
# frozenlist==1.8.0
# fsspec==2025.9.0
# google-ai-generativelanguage==0.6.15
# google-api-core==2.25.2
# google-api-python-client==2.184.0
# google-auth==2.41.1
# google-auth-httplib2==0.2.0
# google-genai==1.41.0
# google-generativeai==0.8.5
# googleapis-common-protos==1.70.0
# greenlet==3.2.4
# grpcio==1.75.1
# grpcio-status==1.71.2
# h11==0.16.0
# httpcore==1.0.9
# httplib2==0.31.0
# httptools==0.6.4
# httpx==0.28.1
# httpx-sse==0.4.1
# huggingface-hub==0.35.3
# humanfriendly==10.0
# idna==3.10
# importlib_metadata==8.7.0
# importlib_resources==6.5.2
# InstructorEmbedding==1.0.1
# Jinja2==3.1.6
# joblib==1.5.2
# jsonpatch==1.33
# jsonpointer==3.0.0
# jsonschema==4.25.1
# jsonschema-specifications==2025.9.1
# kubernetes==34.1.0
# langchain==0.3.27
# langchain-chroma==0.2.6
# langchain-community==0.3.30
# langchain-core==0.3.78
# langchain-google-genai==2.1.12
# langchain-huggingface==0.3.1
# langchain-text-splitters==0.3.11
# langsmith==0.4.32
# markdown-it-py==4.0.0
# MarkupSafe==3.0.3
# marshmallow==3.26.1
# mdurl==0.1.2
# mmh3==5.2.0
# mpmath==1.3.0
# multidict==6.7.0
# mypy_extensions==1.1.0
# networkx==3.5
# numpy==2.3.3
# oauthlib==3.3.1
# onnxruntime==1.23.0
# opentelemetry-api==1.37.0
# opentelemetry-exporter-otlp-proto-common==1.37.0
# opentelemetry-exporter-otlp-proto-grpc==1.37.0
# opentelemetry-proto==1.37.0
# opentelemetry-sdk==1.37.0
# opentelemetry-semantic-conventions==0.58b0
# orjson==3.11.3
# overrides==7.7.0
# packaging==25.0
# pdf2image==1.17.0
# pillow==11.3.0
# posthog==5.4.0
# propcache==0.4.0
# proto-plus==1.26.1
# protobuf==5.29.5
# pyasn1==0.6.1
# pyasn1_modules==0.4.2
# pybase64==1.4.2
# pydantic==2.11.10
# pydantic-settings==2.11.0
# pydantic_core==2.33.2
# Pygments==2.19.2
# PyMuPDF==1.26.4
# pyparsing==3.2.5
# pypdf==6.1.1
# PyPika==0.48.9
# pyproject_hooks==1.2.0
# pyreadline3==3.5.4
# pytesseract==0.3.13
# python-dateutil==2.9.0.post0
# python-dotenv==1.1.1
# PyYAML==6.0.3
# referencing==0.36.2
# regex==2025.9.18
# requests==2.32.5
# requests-oauthlib==2.0.0
# requests-toolbelt==1.0.0
# rich==14.1.0
# rpds-py==0.27.1
# rsa==4.9.1
# safetensors==0.6.2
# scikit-learn==1.7.2
# scipy==1.16.2
# sentence-transformers==5.1.1
# setuptools==80.9.0
# shellingham==1.5.4
# six==1.17.0
# sniffio==1.3.1
# SQLAlchemy==2.0.43
# starlette==0.48.0
# sympy==1.14.0
# tabulate==0.9.0
# tenacity==9.1.2
# threadpoolctl==3.6.0
# tokenizers==0.22.1
# tqdm==4.67.1
# transformers==4.57.0
# typer==0.19.2
# typing-inspect==0.9.0
# typing-inspection==0.4.2
# typing_extensions==4.15.0
# uritemplate==4.2.0
# urllib3==2.3.0
# uvicorn==0.37.0
# watchfiles==1.1.0
# websocket-client==1.8.0
# websockets==15.0.1
# yarl==1.22.0
# zipp==3.23.0
# zstandard==0.25.0


# backend/requirements.txt

# --- Web Framework & Server ---
# Used to build and run the API
fastapi
uvicorn
python-multipart

# --- AI & Language Model ---
# Core components for the RAG pipeline
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-chroma
langchain-google-genai
google-api-python-client
sentence-transformers

# --- Vector Database ---
# Local vector store for embeddings
chromadb

# --- Data Processing & Utilities ---
# For handling PDF, Image, and Text data extraction
pypdf
pytesseract
pdf2image
Pillow
PyMuPDF
requests
beautifulsoup4

# --- Configuration ---
# For managing environment variables like API keys
python-dotenv